NODE-782: Move CLI and DockerCLI to test/cl_node/cli.py
